reducing article components
