+++ src/components/Article.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
+import { graphql } from 'gatsby';
 
-import { useLocale, getContentLanguage } from '~/components/LanguageSwitcher';
-
-import Title, { TitleProps } from './components/Title';
-import Tags, { TagsProps } from './components/Tags';
+import { useLocale, getContentLanguage, Link } from '~/components/LanguageSwitcher';
 
 const Container = styled.article`
   margin: 0 0 60px 0;
@@ -70,16 +68,63 @@ export const Content = styled.section`
   }
 `;
 
+const TagList = styled.ul`
+  display: flex;
+  flex-wrap: wrap;
+  padding: 0;
+  margin: 0;
+`;
+
+const TagItem = styled.li`
+  list-style: none;
+  padding: 0 10px 0 0;
+`;
+
+const TagLink = styled(Link)`
+  text-decoration: none;
+  box-shadow: none;
+  font-size: 16px;
+  color: ${({ theme }) => theme.accentColor};
+  font-weight: 700;
+  text-transform: lowercase;
+
+  &:hover,
+  &:focus {
+    border-bottom: 1px solid ${({ theme }) => theme.accentColor};
+    outline: none;
+  }
+`;
+
+const Title = styled(Link).attrs((props) => ({
+  as: props.to ? Link : 'h2',
+}))`
+  color: ${({ theme }) => theme.titleColor};
+  box-shadow: none;
+  text-decoration: none;
+  font-size: 36px;
+  margin: 0;
+  font-weight: 700;
+
+  &:hover,
+  &:focus {
+    border-bottom: ${({ theme, to }) => to && `1px solid ${theme.titleColor}`};
+    outline: none;
+  }
+`;
+
 type Nullable<T> = T | null;
 
-export type ArticleProps = TitleProps &
-  TagsProps & {
-    readingTime?: number;
-    language?: Nullable<string>;
-    date?: Nullable<string>;
-    content?: Nullable<string>;
-    kind?: Nullable<string>;
-  };
+export type ArticleProps = {
+  readingTime?: number;
+  language?: Nullable<string>;
+  date?: Nullable<string>;
+  content?: Nullable<string>;
+  kind?: Nullable<string>;
+  tags?: string[] | null;
+  title: string;
+  url?: string | null;
+  showContent?: boolean | null;
+};
 
 export default function Article(props: ArticleProps): React.ReactElement {
   const locale = useLocale();
@@ -113,13 +158,21 @@ export default function Article(props: ArticleProps): React.ReactElement {
   return (
     <Container data-testid="article-item">
       <Header>
-        <Title title={props.title} url={props.url} language={props.language} />
+        <Title to={props.url} data-testid="article-header-title" language={props.url ? props.language : undefined}>
+          {props.title}
+        </Title>
         {props.kind === 'articles' && (
           <>
             <MetaAttributes>
               {date} {getReadingTime()} · {getContentLanguage(locale, props.language)}
             </MetaAttributes>
-            <Tags tags={props.tags} />
+            <TagList data-testid="article-header-tags">
+              {props.tags?.map((tag: string, index: number) => (
+                <TagItem key={`${index}-${tag}`} data-testid="article-header-tag">
+                  <TagLink to={`/tags/${tag}`}>{`#${tag}`}</TagLink>
+                </TagItem>
+              ))}
+            </TagList>
           </>
         )}
       </Header>
@@ -127,3 +180,25 @@ export default function Article(props: ArticleProps): React.ReactElement {
     </Container>
   );
 }
+
+export const query = graphql`
+  fragment ArticleInformation on MarkdownRemark {
+    html
+    excerpt
+    fields {
+      collection
+      slug
+      language
+      readingTime {
+        minutes
+      }
+    }
+    frontmatter {
+      date
+      title
+      description
+      tags
+      flags
+    }
+  }
+`;
