+++ src/components/__tests__/Article.test.js
